UPDATE: move plot_imgs to dataset.py
